Implement a global patch for playlog updates
Wii Message Board playlogs are handled by checking and writing to the play_rec.dat file in the data folder of the system menu every minute.

Removing the playlog shouldn't be done in the game itself but in a system menu patch, all you have to do is not create "play_rec.dat" when launching a channel.

What my approach does is skip the updates every minute and nothing more, this way it should still update the message board playlog every time you return to the Wii Menu from the HOME Menu.

This patch can (in theory) be used on every official Wii game.